+++ Kitchen.Api/Commands/CatalogCommands.cs
@@ -1,0 +1,14 @@
+﻿using Kitchen.Api.Domain.Enums;
+
+namespace Kitchen.Api.Commands
+{
+    public record AddToCatalogCommand(
+        string Name,
+        UnitType Unit
+     );
+
+    public record ModifyInCatalogCommand(
+        string Name,
+        UnitType? Unit
+     );
+}