+++ Kitchen.Api/Commands/InventoryCommands.cs
@@ -1,0 +1,17 @@
+﻿using Kitchen.Api.Domain.Enums;
+
+namespace Kitchen.Api.Commands
+{
+    public record AddToStockCommand(
+        string Name,
+        double Amount,
+        StorageLocation Location
+     );
+
+    public record ModifyIngredientCommand(
+        string Name,
+        double? Amount,
+        StorageLocation? Location
+     );
+
+}