+++ Kitchen.Api/Controllers/IngredientDefinitionsController.cs
@@ -1,5 +1,7 @@
-﻿using Kitchen.Api.Models.DTOs;
+﻿using System.Xml.Linq;
+using Kitchen.Api.Commands;
 using Kitchen.Api.Domain.Entities;
+using Kitchen.Api.Models.DTOs;
 using Kitchen.Api.Services;
 using Microsoft.AspNetCore.Mvc;
 
@@ -7,38 +9,47 @@ using Microsoft.AspNetCore.Mvc;
 [Route("api/[controller]")]
 public class IngredientDefinitionsController : ControllerBase
 {
-    private readonly IIngredientCatalogService _ingredientCatalogService;
+    private readonly ICatalogService _catalogService;
 
-    public IngredientDefinitionsController(IIngredientCatalogService ingredientCatalogService)
+    public IngredientDefinitionsController(ICatalogService catalogService)
     {
-        _ingredientCatalogService = ingredientCatalogService;
+        _catalogService = catalogService;
     }
 
     [HttpGet]
-    public IActionResult GetAll() => Ok(_ingredientCatalogService.GetAll());
+    public IActionResult GetAll() => Ok(_catalogService.GetAll());
 
     [HttpPost]
-    public IActionResult Create([FromBody] IngredientDefinitionDto request)
+    public IActionResult Create([FromBody] CreateIngredientDefinitionRequest request)
     {
-        var ingredientDefinition = new IngredientDefinition(request.Name, request.Unit);
+        var command = new AddToCatalogCommand(
+                request.Name,
+                request.Unit
+            );
 
-        _ingredientCatalogService.Add(ingredientDefinition);
-        return CreatedAtAction(nameof(GetAll), new { name = ingredientDefinition.Name }, ingredientDefinition);
+        _catalogService.Add(command);
+
+        return CreatedAtAction(nameof(GetAll), new { name = command.Name }, command);
+
     }
 
     [HttpPut("{name}")]
-    public IActionResult Update(string name, [FromBody] IngredientDefinitionDto request)
+    public IActionResult Update(string name, [FromBody] UpdateIngredientDefinitionRequest request)
     {
-        var ingredient = new IngredientDefinition(name, request.Unit);
+        var command = new ModifyInCatalogCommand(
+            name,
+            request.Unit
+        );
 
-        var success = _ingredientCatalogService.Update(ingredient);
+        var success = _catalogService.Update(command);
+
         return success ? NoContent() : NotFound();
     }
 
     [HttpDelete("{name}")]
     public IActionResult Delete(string name)
     {
-        var success = _ingredientCatalogService.Delete(name);
+        var success = _catalogService.Delete(name);
         return success ? NoContent() : NotFound();
     }
 }