+++ Kitchen.Api/Domain/Entities/IngredientDefinition.cs
@@ -10,16 +10,16 @@ namespace Kitchen.Api.Domain.Entities
         public string Name { get; set; } = string.Empty;
 
         [JsonConverter(typeof(JsonStringEnumConverter))]
-        public UnitType Unit { get; private set; }
+        public UnitType? Unit { get; private set; }
 
         public IngredientDefinition(string name, UnitType unit) { 
             Name = name;
             Unit = unit;
         }
 
-        public void ChangeUnitType(UnitType unit)
+        public void ChangeUnitType(UnitType? unit)
         {
-            if (!Enum.IsDefined(unit) || unit is UnitType.Unspecified)
+            if (!unit.HasValue || unit is UnitType.Unspecified)
             {
                 throw new UnknownUnitTypeException();
             }